# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Oauth2 server.",
        "title": "Gandalf API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9100",
    "paths": {
        "/apps": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "app:me:write"
                        ]
                    }
                ],
                "description": "creates an app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "App"
                ],
                "summary": "Creates a new app",
                "operationId": "app-create",
                "parameters": [
                    {
                        "description": "Creates an app",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.AppCreateData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializers.AppSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/me": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:me:read"
                        ]
                    }
                ],
                "description": "get the user who performs the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Get me",
                "operationId": "me-read",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:me:delete"
                        ]
                    }
                ],
                "description": "deletes the user who perform the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Delete me",
                "operationId": "me-delete",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:me:write"
                        ]
                    }
                ],
                "description": "update me",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Update me",
                "operationId": "me-update",
                "parameters": [
                    {
                        "description": "Updates the user with the given data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.UserUpdateData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.UserSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/apps": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "app:me:read"
                        ]
                    }
                ],
                "description": "Get user's created apps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Get user's created apps",
                "operationId": "me-apps",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cursor's page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "cursor's limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.PaginatedAppsSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/connected-apps": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "app:me:read"
                        ]
                    }
                ],
                "description": "Get user's connected apps",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Get user's connected apps",
                "operationId": "me-connected-apps",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cursor's page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "cursor's limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.PaginatedAppsPublicSerializer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/reset-password": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:me:change-password"
                        ]
                    }
                ],
                "description": "Reset my password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Reset my password",
                "operationId": "me-reset-password",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
        "/me/verify": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "user:me:verify"
                        ]
                    }
                ],
                "description": "Verify me",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Me"
                ],
                "summary": "Verify me",
                "operationId": "me-verify",
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/helpers.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "helpers.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "serializers.AppSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializers.appDataSerializer"
                },
                "type": {
                    "type": "string",
                    "example": "app"
                }
            }
        },
        "serializers.CursorSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializers.cursorDataSerializer"
                },
                "type": {
                    "type": "string",
                    "example": "cursor"
                }
            }
        },
        "serializers.PaginatedAppsPublicSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/serializers.appPublicDataSerializer"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/serializers.paginatedAppsSerializerMeta"
                },
                "type": {
                    "type": "string",
                    "example": "app"
                }
            }
        },
        "serializers.PaginatedAppsSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/serializers.appDataSerializer"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/serializers.paginatedAppsSerializerMeta"
                },
                "type": {
                    "type": "string",
                    "example": "app"
                }
            }
        },
...
        "serializers.UserSerializer": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/serializers.userDataSerializer"
                },
                "type": {
                    "type": "string",
                    "example": "user"
                }
            }
        },
        "serializers.appDataSerializer": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string",
                    "example": "4722679b-5a48-4e85-9084-605e8df610f4"
                },
                "client_secret": {
                    "type": "string",
                    "example": "iuhgf3874tiu34gtwerbguv3iu74"
                },
                "icon_url": {
                    "type": "string",
                    "example": "https://rb.gy/1akgfo"
                },
                "name": {
                    "type": "string",
                    "example": "MyApp"
                },
                "redirect_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://localhost:/callback"
                    ]
                },
                "uuid": {
                    "type": "string",
                    "example": "4722679b-5a48-4e85-9084-605e8df610f4"
                }
            }
        },
        "serializers.appPublicDataSerializer": {
            "type": "object",
            "properties": {
                "icon_url": {
                    "type": "string",
                    "example": "https://rb.gy/1akgfo"
                },
                "name": {
                    "type": "string",
                    "example": "MyApp"
                },
                "uuid": {
                    "type": "string",
                    "example": "4722679b-5a48-4e85-9084-605e8df610f4"
                }
            }
        },
        "serializers.cursorDataSerializer": {
            "type": "object",
            "properties": {
                "actual": {
                    "type": "integer",
                    "example": 4
                },
                "next": {
                    "type": "integer",
                    "example": 6
                },
                "previous": {
                    "type": "integer",
                    "example": 5
                },
                "size": {
                    "type": "integer",
                    "example": 5
                },
                "total_objects": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "serializers.paginatedAppsSerializerMeta": {
            "type": "object",
            "properties": {
                "cursor": {
                    "$ref": "#/definitions/serializers.CursorSerializer"
                }
            }
        },
        "serializers.userDataSerializer": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string",
                    "example": "1997-12-21"
                },
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "name": {
                    "type": "string",
                    "example": "John"
                },
                "phone": {
                    "type": "string",
                    "example": "+34666123456"
                },
                "surname": {
                    "type": "string",
                    "example": "Doe"
                },
                "uuid": {
                    "type": "string",
                    "example": "4722679b-5a48-4e85-9084-605e8df610f4"
                }
            }
        },
        "validators.AppCreateData": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "icon_url": {
                    "type": "string",
                    "example": "http://youriconurl.dev"
                },
                "name": {
                    "type": "string",
                    "example": "MySuperApp"
                },
                "redirect_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://yourredirecturl.dev"
                    ]
                }
            }
        },
...
        "validators.UserUpdateData": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "My@appPassw0rd"
                },
                "phone": {
                    "type": "string",
                    "example": "+34666123456"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://localhost:9100/oauth/login",
            "tokenUrl": "https://localhost:9100/oauth/token",
            "scopes": {
                "app:me:read": " Grants access to read self created apps",
                "app:me:write": " Grants access to write self created apps",
                "user:me:authorized-app": " Grants access an app to get information about the user",
                "user:me:change-password": " Grants access to change self password",
                "user:me:delete": " Grants access to delete self user",
                "user:me:read": " Grants access to read self user",
                "user:me:verify": " Grants access to verify created user",
                "user:me:write": " Grants access to write self user"
            }
        }
    },
...
}